BayesVLM/requirements.txt
transformers~=4.40.0
datasets~=2.19.0
torch~=2.6.0
pillow~=10.3.0
requests~=2.31.0
matplotlib~=3.8.4
laplace-torch
numpy~=1.26.0
lightning~=2.2.3
torchmetrics
wandb
jaxtyping
webdataset
python-dotenv
pandas
dask
dask[dataframe]
scikit-learn
tabulate
tqdm
huggingface_hub
torchaudio==2.6.0
fsspec
s3fs